requirements.txt
moviepy>=1.0.3
google-api-python-client>=2.0.0
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
numpy==1.24.3
librosa>=0.9.0
python-dotenv>=0.19.0
Pillow>=9.0.0
watchdog==3.0.0
eyed3==0.9.7
scipy>=1.7.0
tk>=0.1.0